--- a/core/interface.go
+++ b/core/interface.go
@@ -439,13 +439,15 @@ var _ HasPBDefinitions = (*InterfaceAnnotatedValue)(nil)
 func (iface *InterfaceAnnotatedValue) PBDefinitions(ctx context.Context) ([]*pb.Definition, error) {
 	defs := []*pb.Definition{}
 	objDef := iface.UnderlyingType.TypeDef().AsObject.Value
-	for name, val := range iface.Fields {
-		fieldDef, ok := objDef.FieldByOriginalName(name)
+	for _, field := range objDef.Fields {
+		// TODO: we skip over private fields, we can't convert them anyways (this is a bug)
+		name := field.OriginalName
+		val, ok := iface.Fields[name]
 		if !ok {
-			// TODO: private field; skip. (this is a bug)
+			// missing field
 			continue
 		}
-		fieldType, ok, err := iface.UnderlyingType.SourceMod().ModTypeFor(ctx, fieldDef.TypeDef, true)
+		fieldType, ok, err := iface.UnderlyingType.SourceMod().ModTypeFor(ctx, field.TypeDef, true)
 		if err != nil {
 			return nil, fmt.Errorf("failed to get mod type for field %q: %w", name, err)
 		}