--- a/core/schema/container.go
+++ b/core/schema/container.go
@@ -7,6 +7,7 @@ import (
 	"io/fs"
 	"os"
 	"path"
+	"path/filepath"
 	"slices"
 	"strconv"
 	"strings"
@@ -18,6 +19,7 @@ import (
 	"github.com/moby/buildkit/client/llb/sourceresolver"
 	"github.com/moby/buildkit/exporter/containerimage/exptypes"
 	"github.com/moby/buildkit/frontend/dockerfile/shell"
+	bkgw "github.com/moby/buildkit/frontend/gateway/client"
 	"github.com/moby/buildkit/identity"
 	specs "github.com/opencontainers/image-spec/specs-go/v1"
 	"github.com/vektah/gqlparser/v2/ast"
@@ -566,7 +568,7 @@ func (s *containerSchema) Install() {
 			View(BeforeVersion("v0.12.0")).
 			Extend(),
 
-		dagql.NodeFunc("asTarball", s.asTarball).
+		dagql.NodeFunc("asTarball", DagOpFileMiddleware(s.srv, s.asTarball)).
 			Doc(`Returns a File representing the container serialized to a tarball.`).
 			ArgDoc("platformVariants",
 				`Identifiers for other platform specific containers.`,
@@ -1807,44 +1809,6 @@ func (s *containerSchema) asTarball(
 	parent dagql.Instance[*core.Container],
 	args containerAsTarballArgs,
 ) (inst dagql.Instance[*core.File], err error) {
-	if op := core.DagOpFromContext(ctx); op == nil {
-		fs, err := parent.Self.FSState()
-		if err != nil {
-			return inst, err
-		}
-
-		filename := "/container.tar"
-		st, err := core.DagLLB(ctx, s.srv,
-			[]dagql.Selector{
-				{
-					Field: "directory",
-				},
-				{
-					Field: "withFile",
-					Args: []dagql.NamedInput{
-						{
-							Name:  "path",
-							Value: dagql.String(filename),
-						},
-						{
-							Name:  "source",
-							Value: dagql.NewID[*core.File](dagql.CurrentID(ctx).WithMetadata("", true)),
-						},
-					},
-				},
-			}, []llb.State{fs},
-		)
-		if err != nil {
-			return inst, err
-		}
-
-		f, err := core.NewFileSt(ctx, parent.Self.Query, st, filename, parent.Self.Platform, parent.Self.Services)
-		if err != nil {
-			return inst, err
-		}
-		return dagql.NewInstanceForCurrentID(ctx, s.srv, parent, f)
-	}
-
 	platformVariants, err := dagql.LoadIDs(ctx, s.srv, args.PlatformVariants)
 	if err != nil {
 		return inst, err
@@ -1934,13 +1898,29 @@ func (s *containerSchema) asTarball(
 	defer os.RemoveAll(tmpDir)
 	fileName := identity.NewID() + ".tar"
 
-	def, err := bk.ContainerImageToTarball(ctx, engineHostPlatform.Spec(), tmpDir, fileName, inputByPlatform, opts)
+	err = bk.ContainerImageToTarball(ctx, engineHostPlatform.Spec(), filepath.Join(tmpDir, fileName), inputByPlatform, opts)
 	if err != nil {
 		return inst, fmt.Errorf("container image to tarball file conversion failed: %w", err)
 	}
-	dgst, err := core.GetContentHashFromDef(ctx, bk, def, "/")
+	localDef, err := llb.Local(tmpDir,
+		llb.SessionID(bk.ID()), // see engine/server/bk_session.go, we have a special session that points to our engine host
+		llb.SharedKeyHint(bk.ID()),
+		llb.IncludePatterns([]string{fileName}),
+		llb.WithCustomName(fmt.Sprintf("container-image-to-tarball-%s", fileName)),
+		buildkit.WithTracePropagation(ctx),
+	).Marshal(ctx, llb.Platform(engineHostPlatform.Spec()))
 	if err != nil {
-		return inst, fmt.Errorf("failed to get content hash from definition: %w", err)
+		return inst, fmt.Errorf("failed to create llb definition for container image to tarball: %w", err)
+	}
+	def := localDef.ToPB()
+
+	// force-evaluate to get this definitely into the llb.Local cache
+	_, err = bk.Solve(ctx, bkgw.SolveRequest{
+		Definition: def,
+		Evaluate:   true,
+	})
+	if err != nil {
+		return inst, err
 	}
 
 	fileInst, err := dagql.NewInstanceForCurrentID(ctx, s.srv, parent,
@@ -1949,7 +1929,7 @@ func (s *containerSchema) asTarball(
 	if err != nil {
 		return inst, err
 	}
-	return fileInst.WithMetadata(dgst, true), nil
+	return fileInst, err
 }
 
 type containerImportArgs struct {