Fixes #6927, enables #9098, and continues the project of rebuilding the Ship of Theseus :tada:

This PR introduces a new way of marrying dagger's dagql and buildkit's LLB. Currently, dagql is used to query the graphql API schema, while LLB is used to query the buildkit dag. dagql is much more flexible and easy to extend, while LLB has significantly better caching behavior.

`DagOp` allows us to combine the best of both worlds. Essentially, it allows us to execute a graphql selection *as a buildkit op* - caching it exactly as a native buildkit operation.

## How it works

- We inject custom "LLB" by piggybacking on the `Description` map of an existing LLB operation. Then during `ResolveOp`, we transform this into a magical `CustomOpWrapper`. This LLB is very hacky, but such is life :shrug: Without entirely ripping out LLB (or forking that package), there's not really a neat way of handling this.
- The `CustomOpWrapper` delegates all of it's operations to `Register`ed `CustomOp`s - at the moment, only `DagOp`.
	- Note: since `CustomOp`s are serialized into a `string`, then the `DagOp` has to be JSON-marshallable. This means we have to do some hackery around sending `ID`s instead of `Selector`s directly.
- The `DagOp` evaluates the query, and gets the result as a root `Directory` - it then evaluates this the `Directory`s ref, and returns it.

This effectively allows the dagql query execution to be cached!

## Performance

To show this off, I've ported `AsTarball` to use this (to fix #6927). Essentially, `AsTarball` attempts to self-call itself *inside* a `DagOp` (using a similar technique to #8724). This now means that if the `Container` receiver hasn't changed, then the `File` result isn't forced to be re-evaluated (which it was before).

Just this simple transformation seems to shave a couple seconds off an engine build with `dagger call engine container`.

Without this patch:

```
$ time dagger call engine container
...
Full trace at https://dagger.cloud/jedevc/traces/85dea4974204db7872fb31f732bbdcb2
dagger call engine container  5.77s user 1.31s system 25% cpu 27.576 total
```

With this patch:

```
$ time dagger call engine container
...
Full trace at https://dagger.cloud/jedevc/traces/b384cb5631d72ebaddab576fae164d24
dagger-with-dev dagger call engine container  6.38s user 1.45s system 30% cpu 25.252 total
```

In the traces, you can see that before, the tar unpacking as part of unbundling the SDK images is uncached, while after, they are:

Before:
![image](https://github.com/user-attachments/assets/74d69886-6aec-4dec-a038-f3ab1b5c861c)
After:
![image](https://github.com/user-attachments/assets/980be363-ee54-4159-a259-355a436337d6)

## TODO

- [x] Work out some better serialization for `Selector`s, `ID`s might not be it (#9204 is probably related)
- [x] Ensure we get reasonable purity in `AsTarball`
- [x] Work out how to test this