@jedevc 2025-01-30
Oooookay, have done a lot of refactoring, and now - these DagOps are implemented as middlewares :tada: And all tests are passing :smile: There's also a few related dagql fixes+improvements that I found while in the area.

This is *the first step* - it's very easy to mark the `asTarball` method to be cached using buildkit now, with very very few changes to the actual content of the method - we'll be able to reuse this approach later.

However, the end goal is to completely own the cache refs - this is completely possible, it's just not *immediately* apparent how this should look in dagql. One option is something like https://github.com/dagger/dagger/commit/7a36d890cf96d1a3a21abc85fe5579bb121a7557. This is another custom op, and requires a lot more manual intervention, but you can write directly to a cache ref and skip the nasty `llb.Local` hack. Eventually we can switch to this approach, but didn't want to keep blocking this PR from getting in.