@sipsma 2025-02-26
> Anything specific to do here? I've added the benchmark label, so we can get some data from that hopefully? Not really sure if there's anything specific to do for measuring "leakage".

For now, just start the dev engine locally with debug endpoints exposed, run some reasonable set of tests for a while (I often just run `TestModule` for 5 or so minutes), ctrl-c and then see what the RSS looks like. Hit the `/debug/gc` and pprofs endpoints of the dev engine to double check things.