@sipsma 2025-02-28
> Still seeing a lot of log.newRing (we know about) and FileActionMkFile.Unmarshal (we know about), but at least nothing new seems to be leaking.

@jedevc those shouldn't be leaking anymore after the other fixes, if they are then something might be going wrong. I'll try locally too.